// ts/chat-react-core/gen/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Duration: { input: string; output: string; }
  InputTimestamp: { input: any; output: any; }
  Timestamp: { input: string; output: string; }
};
// ...
export enum ApiEntityState {
  EntityStateBlocked = 'ENTITY_STATE_BLOCKED',
  EntityStateDeleted = 'ENTITY_STATE_DELETED',
  EntityStateUnspecified = 'ENTITY_STATE_UNSPECIFIED'
}
// ...
export type AuthAccount = {
  __typename?: 'AuthAccount';
  acceptedTerms: Array<AuthTermsVersion>;
  birthday?: Maybe<AuthDate>;
  channelCreationEligibility: ChannelChannelCreationEligibility;
  createdAt: Scalars['Timestamp']['output'];
  email: Scalars['String']['output'];
  emailVerifiedAt?: Maybe<Scalars['Timestamp']['output']>;
  externalIds: Array<AuthIdentity>;
  flags: Array<AuthAccountStatusFlag>;
  isBot: Scalars['Boolean']['output'];
  labels: Array<Scalars['String']['output']>;
  marketingConsent: AuthConsentStatus;
  matureRatedContentAllowed: Scalars['Boolean']['output'];
  pendingAgreements: Array<AgreementAgreementRevision>;
  roles: Array<AuthPlatformRole>;
  signupLocation: Scalars['String']['output'];
  signupOrigin?: Maybe<AuthSignupOrigin>;
  staff: Scalars['Boolean']['output'];
  state: ApiEntityState;
  temporary: Scalars['Boolean']['output'];
  uid: Scalars['String']['output'];
};
// ...
export type BadgeBadge = {
  __typename?: 'BadgeBadge';
  level: Scalars['Int']['output'];
  nextLevelAt: Scalars['Timestamp']['output'];
  type: BadgeBadgeType;
};
// ...
export type ChannelAvatarItemPurchase = {
  __typename?: 'ChannelAvatarItemPurchase';
  item: ItemItem;
  itemId: Scalars['ID']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelBundlePurchase = {
  __typename?: 'ChannelBundlePurchase';
  bundleName: Scalars['String']['output'];
  streamerCardIds: Array<Scalars['String']['output']>;
  streamerCards?: Maybe<Array<GameLogicStreamerCard>>;
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};

export type ChannelChallengesFeatureStatus = {
  __typename?: 'ChannelChallengesFeatureStatus';
  enabled: Scalars['Boolean']['output'];
};

/** Public channel information */
export type ChannelChannel = {
  __typename?: 'ChannelChannel';
  activeStream?: Maybe<ChannelStreamSummary>;
  channelFriends: FriendsChannelFriends;
  createdAt: Scalars['Timestamp']['output'];
  currentChatId?: Maybe<Scalars['String']['output']>;
  currentStream?: Maybe<ChannelStream>;
  currentStreamId: Scalars['ID']['output'];
  description: Scalars['String']['output'];
  features: ChannelChannelFeatures;
  followerCount: Scalars['Int']['output'];
  following: Scalars['Boolean']['output'];
  game: GameGame;
  gameId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  isPublic: Scalars['Boolean']['output'];
  itemStats: Array<ItemItemStat>;
  links: Array<ChannelChannelLink>;
  liveStatus: ChannelLiveStatus;
  logo: Scalars['String']['output'];
  matchOngoing: Scalars['Boolean']['output'];
  matureRatedContent: Scalars['Boolean']['output'];
  monetizationSettings: ChannelMonetizationSettings;
  name: Scalars['String']['output'];
  offlineBanner: Scalars['String']['output'];
  offlineBannerUrl: Scalars['String']['output'];
  playedGameIds: Array<Scalars['String']['output']>;
  priority: Scalars['Int']['output'];
  reportingId: Scalars['ID']['output'];
  riskTier: ChannelRiskTier;
  state: ApiEntityState;
  streamedGames?: Maybe<Array<GameGame>>;
  streamer: ProfileProfile;
  streamerId: Scalars['ID']['output'];
  subscriberCount: Scalars['Int']['output'];
  subscription?: Maybe<SubscriptionChannelSubscription>;
  subscriptionConfig?: Maybe<SubscriptionChannelSubscriptionConfig>;
  suspension: ChannelSuspension;
  thumbnail: Scalars['String']['output'];
  title: Scalars['String']['output'];
  userBanStatus: ChannelUserBanStatus;
  viewerCount: Scalars['Int']['output'];
};
// ...
export type ChannelChannelEventContentContentUnion = ChannelAvatarItemPurchase | ChannelBundlePurchase | ChannelGiftSubscriptionPurchase | ChannelStreamerCardPurchase | ChannelSubscriptionPurchase | ChannelSubscriptionRenewal;
// ...
export type ChannelGiftSubscriptionPurchase = {
  __typename?: 'ChannelGiftSubscriptionPurchase';
  recipientUserIds: Array<Scalars['String']['output']>;
  recipients?: Maybe<Array<ProfileProfile>>;
  tier: Scalars['Int']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};
// ...
export type ChannelStreamerCardPurchase = {
  __typename?: 'ChannelStreamerCardPurchase';
  streamerCard: GameLogicStreamerCard;
  streamerCardId: Scalars['ID']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};

export type ChannelStreamingFeatureStatus = {
  __typename?: 'ChannelStreamingFeatureStatus';
  enabled: Scalars['Boolean']['output'];
  suspension?: Maybe<ChannelSuspension>;
};

export type ChannelSubscriptionPurchase = {
  __typename?: 'ChannelSubscriptionPurchase';
  tier: Scalars['Int']['output'];
  user?: Maybe<ProfileProfile>;
  userId: Scalars['ID']['output'];
};

export type ChannelSubscriptionRenewal = {
  __typename?: 'ChannelSubscriptionRenewal';
  tier: Scalars['Int']['output'];
  userId: Scalars['ID']['output'];
};
// ...
export type FriendsFriendshipStatus = {
  __typename?: 'FriendsFriendshipStatus';
  activity?: Maybe<FriendsActivity>;
  lastStatusChange: Scalars['Timestamp']['output'];
  status: FriendsFriendshipStatusStatus;
};
// ...
export type GameLogicCard = {
  __typename?: 'GameLogicCard';
  activeStreamerCard?: Maybe<GameLogicStreamerCard>;
  activeStreamerCards: Array<GameLogicStreamerCard>;
  availableStreamerCards: Array<GameLogicStreamerCard>;
  backImage: Scalars['String']['output'];
  dealingModules: Array<Scalars['String']['output']>;
  description: Scalars['String']['output'];
  failureModules: Array<Scalars['String']['output']>;
  failureTargetValue: Scalars['Int']['output'];
  familyId: Scalars['ID']['output'];
  frontImage: Scalars['String']['output'];
  gameModes: Array<Scalars['String']['output']>;
  icon: Scalars['String']['output'];
  id: Scalars['ID']['output'];
  isAllOrNothing: Scalars['Boolean']['output'];
  isDealtAtStart: Scalars['Boolean']['output'];
  isEnabled: Scalars['Boolean']['output'];
  isMatchCard: Scalars['Boolean']['output'];
  leveling: GameLogicCardLeveling;
  matchCardId: Scalars['Int']['output'];
  name: Scalars['String']['output'];
  pointsMax: Scalars['Int']['output'];
  pointsMin: Scalars['Int']['output'];
  pointsTimeTarget: Scalars['Int']['output'];
  rarity: RarityRarity;
  roleCharacters: Array<Scalars['String']['output']>;
  scoredCounterIds: Array<Scalars['String']['output']>;
  season: GameSeason;
  seasonId: Scalars['ID']['output'];
  sides: Array<Scalars['String']['output']>;
  successModules: Array<Scalars['String']['output']>;
  targetValue: Scalars['Int']['output'];
  targetValues: Array<GameLogicTargetValue>;
  timerDuration: Scalars['Int']['output'];
  unlockLevel: Scalars['Int']['output'];
};
// ...
export type GameLogicStreamerCard = {
  __typename?: 'GameLogicStreamerCard';
  baseCard: GameLogicCard;
  channel: ChannelChannel;
  channelId: Scalars['ID']['output'];
  draft: Scalars['Boolean']['output'];
  facecam: Scalars['String']['output'];
  facecamUrl: Scalars['String']['output'];
  familyId: Scalars['ID']['output'];
  gameId: Scalars['ID']['output'];
  id: Scalars['ID']['output'];
  image: Scalars['String']['output'];
  imageUrl: Scalars['String']['output'];
  name: Scalars['String']['output'];
  saleConfig?: Maybe<StoreV2StreamerCardSaleConfig>;
  video: Scalars['String']['output'];
  videoUrl: Scalars['String']['output'];
};
// ...
export type PlayerStatsPlayerStats = {
  __typename?: 'PlayerStatsPlayerStats';
  adsWatched: Scalars['Int']['output'];
  boosterUsage?: Maybe<PlayerStatsPlayerStatsBoosterUsage>;
  cardBundlesPurchased: Scalars['Int']['output'];
  cardLevelUps: Scalars['Int']['output'];
  cardsPlayed: Scalars['Int']['output'];
  cardsSucceeded: Scalars['Int']['output'];
  currencySpending?: Maybe<PlayerStatsPlayerStatsCurrencySpending>;
  dailyGoalCardsCompleted: Scalars['Int']['output'];
  dailyGoalCardsSet: Scalars['Int']['output'];
  matchesPlayed: Scalars['Int']['output'];
  partyMatchesPlayed: Scalars['Int']['output'];
  shufflesUsed: Scalars['Int']['output'];
  soloMatchesPlayed: Scalars['Int']['output'];
  timePlayed?: Maybe<Scalars['Duration']['output']>;
};
// ...
export enum ProfileColor {
  Color_6Ec9F7 = 'COLOR_6EC9F7',
  Color_56F6C0 = 'COLOR_56F6C0',
  Color_63F655 = 'COLOR_63F655',
  Color_8686F9 = 'COLOR_8686F9',
  ColorB26Afb = 'COLOR_B26AFB',
  ColorC0F656 = 'COLOR_C0F656',
  ColorF6Ce56 = 'COLOR_F6CE56',
  ColorF6F656 = 'COLOR_F6F656',
  ColorF76Ef7 = 'COLOR_F76EF7',
  ColorF69856 = 'COLOR_F69856',
  ColorUnspecified = 'COLOR_UNSPECIFIED'
}
// ...
export type ProfilePlayedGame = {
  __typename?: 'ProfilePlayedGame';
  game: GameGame;
  id: Scalars['ID']['output'];
  lastPlayedAt: Scalars['Timestamp']['output'];
  progression: GameUserProgression;
  season: GameSeason;
  seasonId: Scalars['ID']['output'];
  userId: Scalars['ID']['output'];
};

export enum ProfilePresenceStatus {
  PresenceStatusOffline = 'PRESENCE_STATUS_OFFLINE',
  PresenceStatusOnline = 'PRESENCE_STATUS_ONLINE',
  PresenceStatusUnspecified = 'PRESENCE_STATUS_UNSPECIFIED'
}
// ...
export type ProfileProfile = {
  __typename?: 'ProfileProfile';
  account?: Maybe<AuthAccount>;
  avatarConfig?: Maybe<ProfileProfileAvatarConfig>;
  /**
   * contains a concatenated list of different avatar urls (use avatars object instead)
   * @deprecated field is deprecated
   */
  avatarUrl: Scalars['String']['output'];
  avatars?: Maybe<ProfileProfileAvatars>;
  badges: Array<BadgeBadge>;
  bio: Scalars['String']['output'];
  canChangeUsernameAt?: Maybe<Scalars['Timestamp']['output']>;
  channel?: Maybe<ChannelChannel>;
  discordUsername?: Maybe<Scalars['String']['output']>;
  /** @deprecated field is deprecated */
  displayName: Scalars['String']['output'];
  friends: Array<ProfileProfile>;
  friendshipStatus: FriendsFriendshipStatus;
  isNewUsername: Scalars['Boolean']['output'];
  lastSeen?: Maybe<Scalars['Timestamp']['output']>;
  onlineStatus: ProfilePresenceStatus;
  playedGames: Array<ProfilePlayedGame>;
  preferredColor: ProfileColor;
  settings?: Maybe<ProfileProfileSettings>;
  state: ApiEntityState;
  stats: PlayerStatsPlayerStats;
  temporary: Scalars['Boolean']['output'];
  userId: Scalars['ID']['output'];
  userTag: Scalars['String']['output'];
  usernameHistory?: Maybe<Array<ProfileUsernameChange>>;
  visibility: ProfileProfileVisibility;
};
// ...
export type ProfileProfileAvatarConfig = {
  __typename?: 'ProfileProfileAvatarConfig';
  model: AvatarAvatar;
  modelId: Scalars['ID']['output'];
};

export type ProfileProfileAvatars = {
  __typename?: 'ProfileProfileAvatars';
  avatar2D: Scalars['String']['output'];
  avatar3D: Scalars['String']['output'];
  avatarFullbody: Scalars['String']['output'];
  avatarGender: Scalars['String']['output'];
};

export type ProfileProfileSettings = {
  __typename?: 'ProfileProfileSettings';
  friends: FriendsFriendsSettings;
  privacy: ProfilePrivacySettings;
};

export type ProfileProfileUpdateInput = {
  bio?: InputMaybe<Scalars['String']['input']>;
  discordUsername?: InputMaybe<Scalars['String']['input']>;
  /** @deprecated field is deprecated */
  displayName?: InputMaybe<Scalars['String']['input']>;
  preferredColor?: InputMaybe<ProfileColor>;
  userId?: InputMaybe<Scalars['ID']['input']>;
  userTag?: InputMaybe<Scalars['String']['input']>;
};

export enum ProfileProfileVisibility {
  ProfileVisibilityPrivate = 'PROFILE_VISIBILITY_PRIVATE',
  ProfileVisibilityPublic = 'PROFILE_VISIBILITY_PUBLIC',
  ProfileVisibilityUnspecified = 'PROFILE_VISIBILITY_UNSPECIFIED'
}
// ...
export type ProfileUsernameChange = {
  __typename?: 'ProfileUsernameChange';
  changedAt: Scalars['Timestamp']['output'];
  changedBy: Scalars['String']['output'];
  changer?: Maybe<ProfileProfile>;
  newUsername: Scalars['String']['output'];
  oldUsername: Scalars['String']['output'];
  reason: ModerationViolation;
};
// ...
export type StoreV2StreamerCardSaleConfig = {
  __typename?: 'StoreV2StreamerCardSaleConfig';
  cardId: Scalars['ID']['output'];
  channelId: Scalars['ID']['output'];
  enabled: Scalars['Boolean']['output'];
  excludeFromBundles: Scalars['Boolean']['output'];
  period?: Maybe<StoreV2Period>;
};